adicionar testes

+++ src/main/java/com/example/desafioDevMagroRef/Respositories/QuestionRepository.java
@@ -3,6 +3,6 @@ package com.example.desafioDevMagroRef.Respositories;
 import com.example.desafioDevMagroRef.model.Questions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Questions, Integer> {
+public interface QuestionRepository extends JpaRepository<Questions, Long> {
 
 }
